refactor: int 타입을 Long 타입으로 변환

+++ src/main/java/com/example/onlyone/domain/user/service/UserService.java
@@ -219,7 +219,7 @@ public class UserService {
         // 사용자 지갑 생성 및 웰컴 포인트 100000원 지급
         Wallet wallet = Wallet.builder()
                 .user(user)
-                .postedBalance(100000)
+                .postedBalance(100000L)
                 .build();
         
         walletRepository.save(wallet);
@@ -263,7 +263,7 @@ public class UserService {
 
         // 사용자 지갑 정보 조회
         Optional<Wallet> walletOpt = walletRepository.findByUserWithoutLock(user);
-        Integer balance = walletOpt.map(Wallet::getPostedBalance).orElse(0);
+        Long balance = walletOpt.map(Wallet::getPostedBalance).orElse(0L);
 
         return MyPageResponse.builder()
                 .nickname(user.getNickname())
